# web-navigation-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Web Navigation API",
    "version": "1.0.0",
    "description": "AI-powered web navigation API for autonomous agents. Navigate URLs, extract structured data, perform live search, follow link graphs, run adaptive crawls, handle anti-bot challenges, diff page content, and gate navigation actions.",
    "x-agent-callable": true,
    "x-mcp-compatible": true,
    "x-pricing": {
      "free_tier": {
        "requests_per_day": 100,
        "requests_per_month": 3000
      },
      "pay_per_call": {
        "navigate": 0.005,
        "extract-structured-data": 0.006,
        "live-search": 0.005,
        "follow-links": 0.006,
        "adaptive-crawl": 0.008,
        "anti-bot-handling": 0.004,
        "content-diff": 0.005,
        "browser-session": 0.003,
        "screenshot": 0.003,
        "form-detect": 0.004,
        "submit-form-gated": 0.005,
        "download-assets": 0.004,
        "navigate-and-extract": 0.015,
        "execution-gate": 0.002
      }
    }
  },
  "servers": [
    {
      "url": "https://orbis-apis.onrender.com/web-navigation"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/screenshot": {
      "post": {
        "operationId": "screenshot",
        "summary": "Plan and configure a screenshot capture for a URL",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url",
                  "purpose"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "analysis",
                      "diff",
                      "archival",
                      "evidence",
                      "debug"
                    ]
                  },
                  "selector": {
                    "type": "string"
                  },
                  "full_page": {
                    "type": "boolean"
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "png",
                      "jpeg",
                      "webp"
                    ]
                  },
                  "viewport": {
                    "type": "object",
                    "properties": {
                      "width": {
                        "type": "number"
                      },
                      "height": {
                        "type": "number"
                      }
                    }
                  },
                  "wait_for": {
                    "type": "string"
                  },
                  "delay_ms": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Screenshot capture plan with analysis hints",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "purpose": {
                      "type": "string"
                    },
                    "capture_instructions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "step": {
                            "type": "number"
                          },
                          "action": {
                            "type": "string"
                          },
                          "selector": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "recommended_viewport": {
                      "type": "object",
                      "properties": {
                        "width": {
                          "type": "number"
                        },
                        "height": {
                          "type": "number"
                        }
                      }
                    },
                    "format": {
                      "type": "string"
                    },
                    "full_page": {
                      "type": "boolean"
                    },
                    "estimated_file_size_kb": {
                      "type": "number"
                    },
                    "analysis_hints": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "diff_baseline_recommended": {
                      "type": "boolean"
                    },
                    "confidence_per_section": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "number"
                      }
                    },
                    "recommended_actions_priority_order": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "privacy": {
                      "type": "object",
                      "properties": {
                        "data_stored": {
                          "type": "boolean"
                        },
                        "retention": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing url or purpose"
          }
        }
      }
    },
...
  }
}